daveD - moved method onto Extraction
